Remove old systemdb creation
