+++ engine-cockpit-selenium-test/test/ch/ivyteam/enginecockpit/setup/WebTestWizardSystemDb.java
@@ -1,7 +1,6 @@
 package ch.ivyteam.enginecockpit.setup;
 
 import static ch.ivyteam.enginecockpit.util.EngineCockpitUtil.assertCurrentUrlContains;
-import static ch.ivyteam.enginecockpit.util.EngineCockpitUtil.createOldDb;
 import static ch.ivyteam.enginecockpit.util.EngineCockpitUtil.deleteTempDb;
 import static ch.ivyteam.enginecockpit.util.EngineCockpitUtil.resetConfig;
 import static com.codeborne.selenide.Condition.exist;
@@ -10,7 +9,6 @@ import static com.codeborne.selenide.Condition.visible;
 import static com.codeborne.selenide.Selenide.$;
 
 import org.junit.jupiter.api.AfterEach;
-import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
@@ -20,11 +18,6 @@ import ch.ivyteam.enginecockpit.system.WebTestSystemDb;
 
 @IvyWebTest
 class WebTestWizardSystemDb {
-
-  @BeforeAll
-  static void setup() {
-    createOldDb();
-  }
 
   @BeforeEach
   void beforeEach() {
@@ -53,11 +46,6 @@ class WebTestWizardSystemDb {
   @Test
   void connectionResults() {
     WebTestSystemDb.assertConnectionResults();
-  }
-
-  @Test
-  void oldDbConversionNeeded() {
-    WebTestSystemDb.assertSystemDbConversionDialog();
   }
 
   @Test
