+++ engine-cockpit-test-data/processes/SystemDatabase.p.json
@@ -5,92 +5,18 @@
     "data" : "engine.cockpit.test.data.Data"
   },
   "elements" : [ {
-      "id" : "f0",
-      "type" : "RequestStart",
-      "name" : "createOldDatabase.ivp",
-      "config" : {
-        "signature" : "createOldDatabase"
-      },
-      "visual" : {
-        "at" : { "x" : 96, "y" : 64 },
-        "labelOffset" : { "x" : 18, "y" : 50 }
-      },
-      "connect" : [
-        { "id" : "f4", "to" : "f3" }
-      ]
-    }, {
-      "id" : "f1",
-      "type" : "TaskEnd",
-      "visual" : {
-        "at" : { "x" : 352, "y" : 64 }
-      }
-    }, {
-      "id" : "f3",
-      "type" : "Script",
-      "config" : {
-        "output" : {
-          "code" : [
-            "import ch.ivyteam.enginecockpit.system.database.SystemDatabaseCreator;",
-            "SystemDatabaseCreator.createOldDatabase();"
-          ]
-        }
-      },
-      "visual" : {
-        "at" : { "x" : 224, "y" : 64 }
-      },
-      "connect" : [
-        { "id" : "f2", "to" : "f1" }
-      ]
-    }, {
-      "id" : "f5",
-      "type" : "TaskEnd",
-      "visual" : {
-        "at" : { "x" : 352, "y" : 160 }
-      }
-    }, {
-      "id" : "f6",
-      "type" : "RequestStart",
-      "name" : "deleteOldDatabase.ivp",
-      "config" : {
-        "signature" : "deleteOldDatabase"
-      },
-      "visual" : {
-        "at" : { "x" : 96, "y" : 160 },
-        "labelOffset" : { "x" : 18, "y" : 50 }
-      },
-      "connect" : [
-        { "id" : "f9", "to" : "f7" }
-      ]
-    }, {
-      "id" : "f7",
-      "type" : "Script",
-      "config" : {
-        "output" : {
-          "code" : [
-            "import ch.ivyteam.enginecockpit.system.database.SystemDatabaseCreator;",
-            "SystemDatabaseCreator.deleteOldDatabase();"
-          ]
-        }
-      },
-      "visual" : {
-        "at" : { "x" : 224, "y" : 160 }
-      },
-      "connect" : [
-        { "id" : "f8", "to" : "f5" }
-      ]
-    }, {
       "id" : "f10",
       "type" : "Script",
       "config" : {
         "output" : {
           "code" : [
-            "import ch.ivyteam.enginecockpit.system.database.SystemDatabaseCreator;",
-            "SystemDatabaseCreator.deleteTempDatabase();"
+            "import ch.ivyteam.enginecockpit.system.database.DatabaseCreator;",
+            "DatabaseCreator.deleteTempDatabase();"
           ]
         }
       },
       "visual" : {
-        "at" : { "x" : 224, "y" : 256 }
+        "at" : { "x" : 216, "y" : 72 }
       },
       "connect" : [
         { "id" : "f13", "to" : "f11" }
@@ -99,7 +25,7 @@
       "id" : "f11",
       "type" : "TaskEnd",
       "visual" : {
-        "at" : { "x" : 352, "y" : 256 }
+        "at" : { "x" : 344, "y" : 72 }
       }
     }, {
       "id" : "f12",
@@ -109,7 +35,7 @@
         "signature" : "deleteTempDatabase"
       },
       "visual" : {
-        "at" : { "x" : 96, "y" : 256 },
+        "at" : { "x" : 88, "y" : 72 },
         "labelOffset" : { "x" : 25, "y" : 50 }
       },
       "connect" : [
